requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beartype==0.15.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.2
boto3==1.34.8
botocore==1.34.8
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.7
dataclasses-json==0.6.3
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
fsspec==2023.12.2
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h3==3.7.6
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
ipython==8.19.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-bokeh==3.0.7
jupyterlab-widgets==3.0.9
linkify-it-py==2.0.2
llama-index==0.9.22
llm-app==0.3.6
llvmlite==0.41.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
msal==1.26.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
Office365-REST-Python-Client==2.5.4
openai==1.6.1
packaging==23.2
pandas==2.1.4
panel==1.3.6
param==2.0.1
parso==0.8.3
pathway==0.7.7
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-sat==0.1.8.dev12
pytz==2023.3.post1
pyviz_comms==3.0.0
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
s3transfer==0.10.0
scikit-learn==1.3.2
scipy==1.11.4
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlglot==10.6.1
stack-data==0.6.3
streamlit==1.29.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
webencodings==0.5.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
